GMBX-15 GMBX-25 GMBX-24 GMBX-23 GMBX-16 GMBX-22 GMBX-21 GMBX-19 GMBX-20 GMBX-18 GMBX-17

+++ GameShop/Models/Service/Implementation/ProductService.cs
@@ -69,7 +69,7 @@ namespace GameShop.Models.Service.Implementation
             return await _context.Product.ToListAsync();
         }
 
-        public async Task<Product> GetById(int id)
+        public async Task<Product> GetById(long id)
         {
             var product = await _context.Product.FirstOrDefaultAsync(m => m.Id == id);
